+++ SecurityComplianceCenter/changelog.md
@@ -1,5 +1,13 @@
 ﻿# Changelog
 
+## 1.1.9 (2020-05-13)
+
+- New: Command Enable-SccAuthentication - configures WinRM to allow modern authentication when policies forbid the use of basic authentication.
+- Upd: Import-SccLabelLocalizationXml - added validation for invalid characters among the imported XML.
+- Upd: Set-SccLabelLocalization - added input validation for invalid characters
+- Upd: Set-SccLabelLocalization - adding a new language to either displayname or tooltip will add the default text to the other language if needed. SCC requries languages to have both entries filled if ANY tooltip has been defined.
+- Upd: Connect-SCC - added detection for disabled basic auth in WinRM client
+
 ## 1.1.4 (2020-05-04)
 
 - Fix: Import-SccLabelLocalizationXml - invalid parameter on Stop-PSFFunction causes unexpected errors without proper handling